@@ -1,26 +1,19 @@
 {
   "_args": [
     [
-      {
-        "name": "react-addons-update",
-        "raw": "react-addons-update",
-        "rawSpec": "",
-        "scope": null,
-        "spec": "latest",
-        "type": "tag"
-      },
-      "E:\\Development\\React\\QuestionGrid"
+      "react-addons-update@15.3.1",
+      "C:\\Development\\React\\QuestionGrid"
     ]
   ],
-  "_from": "react-addons-update@latest",
-  "_id": "react-addons-update@15.2.1",
+  "_from": "react-addons-update@15.3.1",
+  "_id": "react-addons-update@15.3.1",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-addons-update",
   "_nodeVersion": "4.4.5",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-addons-update-15.2.1.tgz_1468017064294_0.4678618668112904"
+    "host": "packages-16-east.internal.npmjs.com",
+    "tmp": "tmp/react-addons-update-15.3.1.tgz_1471632609267_0.9007232065778226"
   },
   "_npmUser": {
     "email": "paul@oshannessy.com",
@@ -30,20 +23,20 @@
   "_phantomChildren": {},
   "_requested": {
     "name": "react-addons-update",
-    "raw": "react-addons-update",
-    "rawSpec": "",
+    "raw": "react-addons-update@15.3.1",
+    "rawSpec": "15.3.1",
     "scope": null,
-    "spec": "latest",
-    "type": "tag"
+    "spec": "15.3.1",
+    "type": "version"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/react-addons-update/-/react-addons-update-15.2.1.tgz",
-  "_shasum": "4558e42577e7962c49b6580ddef0e8713487b1ff",
+  "_resolved": "https://registry.npmjs.org/react-addons-update/-/react-addons-update-15.3.1.tgz",
+  "_shasum": "a4efa7dbe62cd07217704f69aff26bf100dc769a",
   "_shrinkwrap": null,
-  "_spec": "react-addons-update",
-  "_where": "E:\\Development\\React\\QuestionGrid",
+  "_spec": "react-addons-update@15.3.1",
+  "_where": "C:\\Development\\React\\QuestionGrid",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
@@ -52,9 +45,15 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "4558e42577e7962c49b6580ddef0e8713487b1ff",
-    "tarball": "https://registry.npmjs.org/react-addons-update/-/react-addons-update-15.2.1.tgz"
+    "shasum": "a4efa7dbe62cd07217704f69aff26bf100dc769a",
+    "tarball": "https://registry.npmjs.org/react-addons-update/-/react-addons-update-15.3.1.tgz"
   },
+  "files": [
+    "LICENSE",
+    "PATENTS",
+    "README.md",
+    "index.js"
+  ],
   "homepage": "https://github.com/facebook/react#readme",
   "keywords": [
     "react",
@@ -63,6 +62,10 @@
   "license": "BSD-3-Clause",
   "main": "index.js",
   "maintainers": [
+    {
+      "email": "opensource+npm@fb.com",
+      "name": "fb"
+    },
     {
       "email": "dan.abramov@gmail.com",
       "name": "gaearon"
@@ -87,7 +90,7 @@
   "name": "react-addons-update",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.2.1"
+    "react": "^15.3.1"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -95,5 +98,5 @@
     "url": "git+https://github.com/facebook/react.git"
   },
   "scripts": {},
-  "version": "15.2.1"
+  "version": "15.3.1"
 }
