@@ -1,5 +1,7 @@
 import React,{Component, PropTypes} from "react";
-import QuestionForm from "./QuestionForm"
+import QuestionForm from "./QuestionForm";
+import AnswerType from '../../constants/AnswerTypeEnum';
+import update from "react-addons-update";
 
 class NewQuestion extends Component{
 
@@ -7,7 +9,7 @@ class NewQuestion extends Component{
       this.setState({
          name:"",
          text:"",
-         answerType:"none",
+         answerType:AnswerType.NONE_SELECTED,
          selectionOptions: []
       });
    }
@@ -17,8 +19,9 @@ class NewQuestion extends Component{
    }
 
    handleSubmit(e){
+      console.log("NewQuestion: handling submit");
       e.preventDefault();
-      this.props.questionCallbacks.addQuestion(this.state);
+      this.props.QuestionCallbacks.addQuestion(this.state);
       this.props.history.pushState(null,"/questions");
    }
 
@@ -26,20 +29,38 @@ class NewQuestion extends Component{
       this.props.history.pushState(null,"/");
    }
 
+   addSelectionOption(optionString) {
+      console.log("addSelectionOption: " + optionString);
+      let nextState = update(this.state.selectionOptions, {$push: [optionString]});
+      this.setState({selectionOptions: nextState});
+   }
+
+   deleteSelectionOption(optionString) {
+      let nextState = update(this.state.selectionOptions, {$unshift: [optionString]});
+      this.setState({selectionOptions: nextState});
+   }
+
    render(){
+
+      let answerTypeCallbacks = {
+         addSelectionOption:this.addSelectionOption.bind(this),
+         deleteSelectionOption:this.addSelectionOption.bind(this)
+   };
+
       return (
          <QuestionForm draftQuestion={this.state}
-                   buttonLabel="Add Question"
-                   handleChange={this.handleChange.bind(this)}
-                   handleSubmit={this.handleSubmit.bind(this)}
-                   handleClose={this.handleClose.bind(this)} />
+                  buttonLabel="Add Question"
+                  handleChange={this.handleChange.bind(this)}
+                  handleSubmit={this.handleSubmit.bind(this)}
+                  handleClose={this.handleClose.bind(this)}
+                 AnswerTypeCallbacks={answerTypeCallbacks}
+         />
       );
    }
 }
 
 NewQuestion.propTypes = {
    QuestionCallbacks: PropTypes.object,
-   AnswerTypeCallbacks: PropTypes.object
 };
 
 export default NewQuestion;
